Fix JSON serialization, progress bar, and cleanup deprecated code
- Fix JSON serialization error by casting ushort to int in metadata dictionaries (XmaParser, ScdaFormat, DdxFormat) - System.Text.Json source generator doesn't support UInt16 in Dictionary<string,object>

- Fix NIF converter progress bar jumping backwards with monotonic progress tracking using Interlocked.CompareExchange

- Add ErrorDialogHelper with Copy to Clipboard functionality for error dialogs

- Remove deprecated ScanDump(byte[]) methods from IDumpScanner interface and implementations (EsmRecordFormat, ScdaFormat)

- Update tests to expect int instead of ushort values in metadata assertions

+++ src/Xbox360MemoryCarver/Core/Formats/Ddx/DdxFormat.cs
@@ -149,7 +149,7 @@ public sealed class DdxFormat : FileFormatBase, IFileConverter
 
         var metadata = new Dictionary<string, object>
         {
-            ["version"] = version,
+            ["version"] = (int)version,
             ["gpuFormat"] = formatByte,
             ["isTiled"] = isTiled,
             ["dataOffset"] = 0x44,
